actualizacion de paquetes y documentacion
