+++ nbproject/private/private.xml
@@ -8,6 +8,7 @@
             <file>file:/C:/Users/javie/Desktop/5TO%20SEMESTRE/arquitectura/proyecto/Patolli_equipo03/src/pruebagraficos/Ficha.java</file>
             <file>file:/C:/Users/javie/Desktop/5TO%20SEMESTRE/arquitectura/proyecto/Patolli_equipo03/src/pruebagraficos/CasillaCentral.java</file>
             <file>file:/C:/Users/javie/Desktop/5TO%20SEMESTRE/arquitectura/proyecto/Patolli_equipo03/src/pruebagraficos/Colores.java</file>
+            <file>file:/C:/Users/javie/Desktop/5TO%20SEMESTRE/arquitectura/proyecto/Patolli_equipo03/src/pruebagraficos/Casilla.java</file>
             <file>file:/C:/Users/javie/Desktop/5TO%20SEMESTRE/arquitectura/proyecto/Patolli_equipo03/src/pruebagraficos/Tablero.java</file>
             <file>file:/C:/Users/javie/Desktop/5TO%20SEMESTRE/arquitectura/proyecto/Patolli_equipo03/src/pruebagraficos/Jugador.java</file>
         </group>
