+++ src/Dominio/CnvDibujo.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package pruebagraficos;
+package Dominio;
 
 import java.awt.BasicStroke;
 import java.awt.Canvas;
