+++ src/Dominio/iCasilla.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package pruebagraficos;
+package Dominio;
 
+import Dominio.Ficha;
 import java.awt.Canvas;
 import java.awt.Graphics;
