--- src/pruebagraficos/CasillaTriangular.java
@@ -1,30 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package pruebagraficos;
-
-import java.awt.Graphics;
-
-/**
- *
- * @author javieextends
- */
-public class CasillaTriangular extends iCasilla{
-
-    public CasillaTriangular(int x, int y) {
-        super(x, y);
-    }
-
-    @Override
-    public void analizarJugada() {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public void paint(Graphics g) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-    
-}
